aerial-detector/requirements.txt
ultralytics>=8.2.0
opencv-python>=4.8
numpy>=1.24
pillow>=10.0
pyyaml>=6.0
tqdm>=4.66
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
